fix: correct import paths in consolidated test files
- Fix relative import paths after moving tests to test/ directory
- Update all require() statements to use correct ../../src/ paths
- Fix paths for: unified-rule-processor, shared-validator, board-rules, log
- Maintain test functionality while fixing import resolution
